Update main.c
Shortened the height of the new columns from 18 to 16 (window hides 2 rows). Obstacles now disappear when hit by the plane (had to create a bkg_map (16x32) to hold all map data). Added a score that increments after a full screen scroll. Added function to convert a uin16 score value into 5 tiles for display purposes but it is not connected to the Window tiles yet.

This has a bug that corrupts the background tile map. My suspicion is that the set_bkg_tiles() function is being executed at a point during the display draw that causes corruption. This is an ongoing bug from previous commits but now seem to happen more often (I am doing a lot more calls to that function now).