Changed the way collisions work
The collision map used to contain an identifier of the type of collision (block, wall, pickup). Now, it contains the background element's health, which makes everything destroyable now.

The collision logic now subtracts a specified amount from this health on every collision and only removes the block from the background if the health < 0. Bullets take off 1 health and the plane itself 2. I'm not sure if we need the plane to take off any health but it's there in case it becomes useful.

Walls have 127 health (essentially indestructible) and blocks have 3 health.

The pickups are exempt from this collision logic. Bullets do not collide with pickups and the plane always removes the background block ("picks up " the item).

Increased the maximum number of bullets since it takes more of them to break a block. Increased the number of bombs.