Refactor and simplify player and collision code
Moving the player code to a player.h file helps with the readability of main.c.

From testing, I found that the changes to the collision code actually makes the code run faster. The screen tearing seems to be farther up the screen, which I think means that the code is closer to running completely within the vblank period.