Copy less background map data to VRAM
This is a performance optimization where we only copy the background map data in the visible parts of the screen (plus the 2 upcoming columns) to VRAM. From testing, unfortunately this change does **not** seem to have a big impact on the performance (i.e. getting rid of the screen tear).

Also in this commit: Sort and rearrange the header imports in main.c.