Fix high score external RAM initialization
Checking if the first two bytes of external RAM equal 0xFFFF if not initialized is an invalid assumption, as this is not true on all platforms. This commit fixes this by writing a constant signature into external RAM on initialization and checking that that signature exists in subsequent game runs.