Increase player speed over time
This is to accommodate the fact that the background scroll speed also increases over time. The player's speed is intentionally less than (or equal to) the background scroll speed. This is because 1) it's hard to control the player at high speeds, and 2) moving diagonally essentially doubles the player's speed, which is super uncontrollable at high speeds.

I tested some speed values out, and these values felt good to play.