Initialize random number generator properly
As described in https://gbdk-2020.github.io/gbdk-2020/docs/api/rand_8h.html, I use two non-deterministic times to poll the 8-bit DIV_REG register to get a random 16-bit integer with which to seed the random number generator. After this commit, we truly start on a random biome now, which is fun!